Added fartometer
Added fartometer, but it has some weird problems due to stretching the
images because I was trying to get it to always work on all resolutions.

Messed with the camera follow script settings in the inspector to get it
to follow the player.

Updated the project's editor settings so asset serialization is set to
force text. That's supposed to help with merging.